backend/requirements.txt
# Dependencies
requests>=2.31.0
nltk>=3.8.0
libsql-client>=0.6.0
python-dotenv>=1.0.0
